using generic clear method to clear data from provider in batch test file
